# libs/api-client/src/generated/apis/ProjectsApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime';
import {
    type CreateProjectDto,
    CreateProjectDtoFromJSON,
    CreateProjectDtoToJSON,
} from '../models/CreateProjectDto';
import {
    type ProjectDto,
    ProjectDtoFromJSON,
    ProjectDtoToJSON,
} from '../models/ProjectDto';

export interface DeleteProjectRequest {
    id: string;
}

export interface PostProjectRequest {
    createProjectDto: CreateProjectDto;
}

/**
 * 
 */
export class ProjectsApi extends runtime.BaseAPI {

    /**
     * Creates request options for deleteProject without sending the request
     */
    async deleteProjectRequestOpts(requestParameters: DeleteProjectRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['id'] == null) {
            throw new runtime.RequiredError(
                'id',
                'Required parameter "id" was null or undefined when calling deleteProject().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};


        let urlPath = `/projects/{id}`;
        urlPath = urlPath.replace('{id}', encodeURIComponent(String(requestParameters['id'])));

        return {
            path: urlPath,
            method: 'DELETE',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     */
    async deleteProjectRaw(requestParameters: DeleteProjectRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<void>> {
        const requestOptions = await this.deleteProjectRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.VoidApiResponse(response);
    }

    /**
     */
    async deleteProject(id: string, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<void> {
        await this.deleteProjectRaw({ id: id }, initOverrides);
    }

    /**
     * Creates request options for getProjects without sending the request
     */
    async getProjectsRequestOpts(): Promise<runtime.RequestOpts> {
        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};


        let urlPath = `/projects`;

        return {
            path: urlPath,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        };
    }

    /**
     */
    async getProjectsRaw(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<Array<ProjectDto>>> {
        const requestOptions = await this.getProjectsRequestOpts();
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => jsonValue.map(ProjectDtoFromJSON));
    }

    /**
     */
    async getProjects(initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<Array<ProjectDto>> {
        const response = await this.getProjectsRaw(initOverrides);
        return await response.value();
    }

    /**
     * Creates request options for postProject without sending the request
     */
    async postProjectRequestOpts(requestParameters: PostProjectRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['createProjectDto'] == null) {
            throw new runtime.RequiredError(
                'createProjectDto',
                'Required parameter "createProjectDto" was null or undefined when calling postProject().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';


        let urlPath = `/projects`;

        return {
            path: urlPath,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: CreateProjectDtoToJSON(requestParameters['createProjectDto']),
        };
    }

    /**
     */
    async postProjectRaw(requestParameters: PostProjectRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<ProjectDto>> {
        const requestOptions = await this.postProjectRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => ProjectDtoFromJSON(jsonValue));
    }

    /**
     */
    async postProject(createProjectDto: CreateProjectDto, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<ProjectDto> {
        const response = await this.postProjectRaw({ createProjectDto: createProjectDto }, initOverrides);
        return await response.value();
    }

}