# libs/api-client/src/generated/models/CreateKnowledgeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateKnowledgeDto
 */
export interface CreateKnowledgeDto {
    /**
     * 
     * @type {string}
     * @memberof CreateKnowledgeDto
     */
    text: string;
    /**
     * 
     * @type {string}
     * @memberof CreateKnowledgeDto
     */
    projectId?: string;
}

/**
 * Check if a given object implements the CreateKnowledgeDto interface.
 */
export function instanceOfCreateKnowledgeDto(value: object): value is CreateKnowledgeDto {
    if (!('text' in value) || value['text'] === undefined) return false;
    return true;
}

export function CreateKnowledgeDtoFromJSON(json: any): CreateKnowledgeDto {
    return CreateKnowledgeDtoFromJSONTyped(json, false);
}

export function CreateKnowledgeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateKnowledgeDto {
    if (json == null) {
        return json;
    }
    return {
        
        'text': json['text'],
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
    };
}

export function CreateKnowledgeDtoToJSON(json: any): CreateKnowledgeDto {
    return CreateKnowledgeDtoToJSONTyped(json, false);
}

export function CreateKnowledgeDtoToJSONTyped(value?: CreateKnowledgeDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'text': value['text'],
        'projectId': value['projectId'],
    };
}