# libs/api-client/src/generated/models/CreateProjectDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateProjectDto
 */
export interface CreateProjectDto {
    /**
     * 
     * @type {string}
     * @memberof CreateProjectDto
     */
    name: string;
}

/**
 * Check if a given object implements the CreateProjectDto interface.
 */
export function instanceOfCreateProjectDto(value: object): value is CreateProjectDto {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateProjectDtoFromJSON(json: any): CreateProjectDto {
    return CreateProjectDtoFromJSONTyped(json, false);
}

export function CreateProjectDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateProjectDto {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
    };
}

export function CreateProjectDtoToJSON(json: any): CreateProjectDto {
    return CreateProjectDtoToJSONTyped(json, false);
}

export function CreateProjectDtoToJSONTyped(value?: CreateProjectDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
    };
}