# libs/api-client/src/generated/models/KnowledgeEntryDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface KnowledgeEntryDto
 */
export interface KnowledgeEntryDto {
    /**
     * 
     * @type {string}
     * @memberof KnowledgeEntryDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeEntryDto
     */
    rawText: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof KnowledgeEntryDto
     */
    tags: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof KnowledgeEntryDto
     */
    projectId?: string | null;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeEntryDto
     */
    createdAt: number;
}

/**
 * Check if a given object implements the KnowledgeEntryDto interface.
 */
export function instanceOfKnowledgeEntryDto(value: object): value is KnowledgeEntryDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('rawText' in value) || value['rawText'] === undefined) return false;
    if (!('tags' in value) || value['tags'] === undefined) return false;
    if (!('createdAt' in value) || value['createdAt'] === undefined) return false;
    return true;
}

export function KnowledgeEntryDtoFromJSON(json: any): KnowledgeEntryDto {
    return KnowledgeEntryDtoFromJSONTyped(json, false);
}

export function KnowledgeEntryDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): KnowledgeEntryDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'rawText': json['rawText'],
        'tags': json['tags'],
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
        'createdAt': json['createdAt'],
    };
}

export function KnowledgeEntryDtoToJSON(json: any): KnowledgeEntryDto {
    return KnowledgeEntryDtoToJSONTyped(json, false);
}

export function KnowledgeEntryDtoToJSONTyped(value?: KnowledgeEntryDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'rawText': value['rawText'],
        'tags': value['tags'],
        'projectId': value['projectId'],
        'createdAt': value['createdAt'],
    };
}