# libs/api-client/src/generated/models/KnowledgeQueueItemDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface KnowledgeQueueItemDto
 */
export interface KnowledgeQueueItemDto {
    /**
     * 
     * @type {string}
     * @memberof KnowledgeQueueItemDto
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof KnowledgeQueueItemDto
     */
    position: number;
}

/**
 * Check if a given object implements the KnowledgeQueueItemDto interface.
 */
export function instanceOfKnowledgeQueueItemDto(value: object): value is KnowledgeQueueItemDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('position' in value) || value['position'] === undefined) return false;
    return true;
}

export function KnowledgeQueueItemDtoFromJSON(json: any): KnowledgeQueueItemDto {
    return KnowledgeQueueItemDtoFromJSONTyped(json, false);
}

export function KnowledgeQueueItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): KnowledgeQueueItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'position': json['position'],
    };
}

export function KnowledgeQueueItemDtoToJSON(json: any): KnowledgeQueueItemDto {
    return KnowledgeQueueItemDtoToJSONTyped(json, false);
}

export function KnowledgeQueueItemDtoToJSONTyped(value?: KnowledgeQueueItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'position': value['position'],
    };
}