# libs/api-client/src/generated/models/PlanItemDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PlanItemDto
 */
export interface PlanItemDto {
    /**
     * 
     * @type {string}
     * @memberof PlanItemDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof PlanItemDto
     */
    text: string;
    /**
     * 
     * @type {boolean}
     * @memberof PlanItemDto
     */
    completed: boolean;
    /**
     * 
     * @type {number}
     * @memberof PlanItemDto
     */
    sortOrder: number;
    /**
     * 
     * @type {string}
     * @memberof PlanItemDto
     */
    parentId: string | null;
}

/**
 * Check if a given object implements the PlanItemDto interface.
 */
export function instanceOfPlanItemDto(value: object): value is PlanItemDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('text' in value) || value['text'] === undefined) return false;
    if (!('completed' in value) || value['completed'] === undefined) return false;
    if (!('sortOrder' in value) || value['sortOrder'] === undefined) return false;
    if (!('parentId' in value) || value['parentId'] === undefined) return false;
    return true;
}

export function PlanItemDtoFromJSON(json: any): PlanItemDto {
    return PlanItemDtoFromJSONTyped(json, false);
}

export function PlanItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanItemDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'text': json['text'],
        'completed': json['completed'],
        'sortOrder': json['sortOrder'],
        'parentId': json['parentId'],
    };
}

export function PlanItemDtoToJSON(json: any): PlanItemDto {
    return PlanItemDtoToJSONTyped(json, false);
}

export function PlanItemDtoToJSONTyped(value?: PlanItemDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'text': value['text'],
        'completed': value['completed'],
        'sortOrder': value['sortOrder'],
        'parentId': value['parentId'],
    };
}