# libs/api-client/src/generated/models/PlanItemInputDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PlanItemInputDto
 */
export interface PlanItemInputDto {
    /**
     * 
     * @type {string}
     * @memberof PlanItemInputDto
     */
    text: string;
    /**
     * 
     * @type {boolean}
     * @memberof PlanItemInputDto
     */
    completed: boolean;
    /**
     * 
     * @type {number}
     * @memberof PlanItemInputDto
     */
    depth?: number;
}

/**
 * Check if a given object implements the PlanItemInputDto interface.
 */
export function instanceOfPlanItemInputDto(value: object): value is PlanItemInputDto {
    if (!('text' in value) || value['text'] === undefined) return false;
    if (!('completed' in value) || value['completed'] === undefined) return false;
    return true;
}

export function PlanItemInputDtoFromJSON(json: any): PlanItemInputDto {
    return PlanItemInputDtoFromJSONTyped(json, false);
}

export function PlanItemInputDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanItemInputDto {
    if (json == null) {
        return json;
    }
    return {
        
        'text': json['text'],
        'completed': json['completed'],
        'depth': json['depth'] == null ? undefined : json['depth'],
    };
}

export function PlanItemInputDtoToJSON(json: any): PlanItemInputDto {
    return PlanItemInputDtoToJSONTyped(json, false);
}

export function PlanItemInputDtoToJSONTyped(value?: PlanItemInputDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'text': value['text'],
        'completed': value['completed'],
        'depth': value['depth'],
    };
}