# libs/api-client/src/generated/models/UpdateKnowledgeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Life Planner API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateKnowledgeDto
 */
export interface UpdateKnowledgeDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateKnowledgeDto
     */
    text?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateKnowledgeDto
     */
    projectId?: string | null;
}

/**
 * Check if a given object implements the UpdateKnowledgeDto interface.
 */
export function instanceOfUpdateKnowledgeDto(value: object): value is UpdateKnowledgeDto {
    return true;
}

export function UpdateKnowledgeDtoFromJSON(json: any): UpdateKnowledgeDto {
    return UpdateKnowledgeDtoFromJSONTyped(json, false);
}

export function UpdateKnowledgeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateKnowledgeDto {
    if (json == null) {
        return json;
    }
    return {
        
        'text': json['text'] == null ? undefined : json['text'],
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
    };
}

export function UpdateKnowledgeDtoToJSON(json: any): UpdateKnowledgeDto {
    return UpdateKnowledgeDtoToJSONTyped(json, false);
}

export function UpdateKnowledgeDtoToJSONTyped(value?: UpdateKnowledgeDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'text': value['text'],
        'projectId': value['projectId'],
    };
}